Split iast file so each functionality is on its own file
